chore: Do not drop old table on migration
To allow zero downtime, do not delete the old constance table when we
transition. This allows both to run in parallel and we can manually
delete the table later.

Also handle the case where code loading begins populating the constance
DB on its own and duplicate entries are attempted to be inserted using IODKU.